refactor: 💡 change import style for better clarity
