+++ src/main.test.js
@@ -1,9 +1,9 @@
 const { main } = require('./main')
-const { startNextDevServer } = require('./next-dev-server')
-const { createProgram } = require('./program')
+const startNextDevServer = require('./next-dev-server/start')
+const createProgram = require('./program/create')
 
-jest.mock('./next-dev-server', () => ({ startNextDevServer: jest.fn() }))
-jest.mock('./program', () => ({ createProgram: jest.fn() }))
+jest.mock('./next-dev-server/start.js', () => jest.fn())
+jest.mock('./program/create.js', () => jest.fn())
 
 const mockProgram = jest.fn()
