--- src/next-dev-server/index.js
@@ -1,4 +1,0 @@
-module.exports = {
-  createNextDevServer: require('./create'),
-  startNextDevServer: require('./start'),
-}
