--- src/program/index.js
@@ -1,3 +1,0 @@
-module.exports = {
-  createProgram: require('./create'),
-}
