--- src/server/index.js
@@ -1,3 +1,0 @@
-module.exports = {
-  startServer: require('./start'),
-}
