--- src/watcher/index.js
@@ -1,3 +1,0 @@
-module.exports = {
-  startWatcher: require('./start'),
-}
